age

Function decrypt

pub fn decrypt(
    identity: &impl Identity,
    ciphertext: &[u8],
) -> Result<Vec<u8>, DecryptError>

Decrypts the given ciphertext with the given identity.

If the armor feature flag is enabled, this will also handle armored age ciphertexts.

To attempt decryption with more than one identity, use Decryptor (as well as ArmoredReader if the armor feature flag is enabled).