age::scrypt

Struct Recipient

pub struct Recipient { /* private fields */ }

A passphrase-based recipient. Anyone with the passphrase can decrypt the file.

If an scrypt::Recipient is used, it must be the only recipient for the file: it can’t be mixed with other recipient types and can’t be used multiple times for the same file.

This API should only be used with a passphrase that was provided by (or generated for) a human. For programmatic use cases, instead generate an x25519::Identity.

Implementations

impl Recipient

pub fn new(passphrase: SecretString) -> Self

Constructs a new Recipient with the given passphrase.

The scrypt work factor is picked to target about 1 second for encryption or decryption on this device. Override it with Self::set_work_factor.

pub fn set_work_factor(&mut self, log_n: u8)

Sets the scrypt work factor to N = 2^log_n.

This method must be called before Self::wrap_file_key to have an effect.

Panics

Panics if log_n == 0 or log_n >= 64.