age

Module ssh

The “ssh-rsa” and “ssh-ed25519” recipient types, which allow reusing existing SSH keys for encryption with age-encryption.org/v1.

These recipient types should only be used for compatibility with existing keys, and native X25519 keys should be preferred otherwise.

Note that these recipient types are not anonymous: the encrypted message will include a short 32-bit ID of the public key.

Structs

EncryptedKey

An encrypted SSH private key.

Enums

Identity

An SSH private key for decrypting an age file.

ParseRecipientKeyError

Error conditions when parsing an SSH recipient.

Recipient

A key that can be used to encrypt a file to a recipient.

UnsupportedKey

A key that we know how to parse, but that we do not support.